// pr_ciudadano_frontend/src/app/components/grafica-congresal-candidato/grafica-congresal-candidato.interfaces.ts
export interface IBarInfo {
  number_of_valid_votes: number;
  number_of_candidate?: string;
  name_of_candidate: string;
  url_candidate_image: string;
  percentage_of_valid_votes?: number;
  urlAgrupacionImage: string;
  number_of_list?: number;
  code_of_politic_group?: string;
  name_of_politic_group?: string;
  group?: number;
}