// pr_ciudadano_frontend/src/app/helpers/estado-enum.ts
export enum EstadoActa {
  PENDIENTE = "PEN",
  OBSERVADA = "OBS",
  CONTABILIZADA = "CON",
}

export enum EnumAmbito {
  NACIONAL = "NACIONAL",
  EXTRANJERO = "EXTRANJERO",
}

export enum EnumMensajeGeneral {
  MENSAJE_PRESIONE_BOTON_FILTRAR = "Para continuar, por favor presione el botón FILTRAR.",
  MENSAJE_NO_SE_ENCONTRARON_RESULTADOS = "Lo sentimos, no se encontraron resultados que coincidan con su búsqueda. Por favor, verifique los criterios ingresados e inténtelo nuevamente.",
  MENSAJE_INFORMACION_SELECCIONADA = "Todavía no se cuenta con información para mostrar en la opción seleccionada.",
}